openai classes called from ottollm, generate-prompt method shifted to utils
