+++ django/chat/urls.py
@@ -110,6 +110,5 @@ urlpatterns = [
         views.set_preset_default,
         name="set_preset_default",
     ),
-    path("reset-form/", views.reset_form_view, name="reset_form_view"),
     path("generate-prompt/", views.generate_prompt_view, name="generate_prompt_view"),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
